requirements.txt
iai-gxipy
matplotlib
opencv-python
numpy

